Remove Skyscape Interim AnyConnect profile
Since moving to Platform 1, we no longer need a VPN to connect to vCloud
Director. Instead, we authenticate through the Skyscape portal, which then
passes through to vCloud Director. As a result, Boxen no longer needs to
configure the Skyscape Interim platform AnyConnect profile as we do not need
to use it.