Update to the latest ruby module
There are some breakages between this version and the previous one:
https://github.com/boxen/puppet-ruby#breakages-since-last-major-version

We don't use `rbenv::plugin` or `chruby` and we already have the
`module-data` module, so those aren't problems. I've installed `bundler`
for all rubies, so that's dealt with.

The moving of rubies from /opt/boxen/rbenv/versions to /opt/rubies makes
this a bit of a warm update, as it's going to build new ruby versions.
It's best to get lunch or a coffee while boxen is running.