Manage printers - Secure Printer 4
Automatically setup "Secure Printer 4" which is on the 6th floor of Aviation
House. This saves enabling the CUPS web interface, finding out the IP and
queue name, and adding it by hand, which is what I did a long time ago.

This resource is not idempotent because `lpstat -v` doesn't report the
username in the LPD URI. I can live with this though, for not having to know
about CUPS again.