Use specific printer driver
Using the preferred printer driver should give better printer output,
and allow printer-specific options to be used when configuring.
This change gives more control to the end user.

Note: the driver may need to be installed via munki, in which case we
should not merge this. Not everyone currently has a managed machine.