Vendor version 1.8.2 of json_pure gem
Vendor version 1.8.2 of the `json_pure` gem, which was added in 732900c0
but then removed by 97e4e1e8.

It was also missed in 37ff67e43e2, which changed the version of
`json_pure` in `Gemfile.lock` from 1.8.1 to 1.8.2.

The changes in this commit are the result of running `bundle install`.